tests/wpt/web-platform-tests/html/browsers/sandboxing/sandbox-disallow-same-origin.html

<!DOCTYPE html>
<html>
  <head>
    <title>Access to sandbox iframe</title>
    <link rel="author" title="Kinuko Yasuda" href="mailto:kinuko@chromium.org">
    <link rel="help" href="http://www.w3.org/html/wg/drafts/html/master/browsers.html#sandboxing">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>

  <body>
    <h1>Access to sandbox iframe</h1>
    <script type="text/javascript">
      var t = async_test("Access to sandbox iframe is disallowed")
      var called = 0;
      function calledFromIframe() {
        called++;
      }
      function loaded() {
        assert_equals(document.getElementById('sandboxedframe').contentWindow.document, undefined);
        assert_equals(called, 0);
        t.done();
      }
    </script>

    <iframe src="/html/browsers/sandboxing/inner-iframe.html" style="visibility:hidden;display:none" sandbox id="sandboxedframe" onload="loaded();"></iframe>
  </body>

    <div id="log"></div>
</html>
